+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased (MVP)
 
 ## 07/03/2021
+-   Add `take` and `skip` query params to `/users/all` route.
 -   Add `version_checker` route.
 -   Change `ResponseBody` struct.
 -   Remove password from `UserModel` JSON serialize.